modify : 리뷰 기능 및 css 업데이트

+++ src/components/Review/myReview.js
@@ -2,16 +2,19 @@ import React, { useState, useEffect } from 'react';
 import css from './myReview.module.scss';
 import { FaStar } from 'react-icons/fa';
 
-const deleteReview = async () => {
-  const res = await fetch('http://localhost:10010/review/4', {
-    method: 'DELETE',
-  });
-  const json = await res.json();
+function MyReview({ data, reviewOnClick }) {
+  const onDelete = async () => {
+    // alert('정말 리뷰를 삭제하시겠습니까?');
+    if (window.confirm('정말 리뷰를 삭제하시겠습니까?')) {
+      const res = await fetch('http://localhost:10010/reservation/toReview/2', {
+        method: 'DELETE',
+      });
+      const json = await res.json();
+      console.log(json);
+    }
+  };
 
-  console.log(json);
-};
 
-function MyReview({ data, reviewOn }) {
   return (
     <div className={css.container}>
       <div className={css.myReview}>
@@ -27,8 +30,10 @@ function MyReview({ data, reviewOn }) {
       </div>
 
       <div className={css.button_group}>
-        <button onClick={reviewOn}>수정</button>
-        <button onClick={deleteReview}>삭제</button>
+
+        <button onClick={() => reviewOnClick(data.idx)}>수정</button>
+        <button onClick={onDelete}>삭제</button>
+
       </div>
     </div>
   );
